style clean up; organize fns and clean up
